feat(clerk-js): Improve Card element, cleanup parts of UI
