+++ packages/clerk-js/src/ui/common/CustomPageContentContainer.tsx
@@ -1,5 +1,5 @@
 import { Col, descriptors } from '../customizables';
-import { CardAlert, NavbarMenuButtonRow, useCardState, withCardStateProvider } from '../elements';
+import { Card, NavbarMenuButtonRow, useCardState, withCardStateProvider } from '../elements';
 import type { CustomPageContent } from '../utils';
 import { ExternalElementMounter } from '../utils';
 
@@ -11,7 +11,7 @@ export const CustomPageContentContainer = withCardStateProvider(
         elementDescriptor={descriptors.page}
         gap={8}
       >
-        <CardAlert>{card.error}</CardAlert>
+        <Card.Alert>{card.error}</Card.Alert>
         <NavbarMenuButtonRow />
         <Col
           elementDescriptor={descriptors.profilePage}
