+++ packages/clerk-js/src/ui/components/OrganizationProfile/OrganizationProfile.tsx
@@ -34,13 +34,15 @@ const _OrganizationProfile = (_: OrganizationProfileProps) => {
 const AuthenticatedRoutes = withCoreUserGuard(() => {
   const contentRef = React.useRef<HTMLDivElement>(null);
   return (
-    <ProfileCard
+    <ProfileCard.Root
       sx={t => ({ display: 'grid', gridTemplateColumns: '1fr 3fr', height: t.sizes.$176, overflow: 'hidden' })}
     >
       <OrganizationProfileNavbar contentRef={contentRef}>
-        <OrganizationProfileRoutes contentRef={contentRef} />
+        <ProfileCard.Content contentRef={contentRef}>
+          <OrganizationProfileRoutes />
+        </ProfileCard.Content>
       </OrganizationProfileNavbar>
-    </ProfileCard>
+    </ProfileCard.Root>
   );
 });
