+++ packages/clerk-js/src/ui/components/UserProfile/UserProfileRoutes.tsx
@@ -1,9 +1,7 @@
 import { CustomPageContentContainer } from '../../common/CustomPageContentContainer';
 import { USER_PROFILE_NAVBAR_ROUTE_ID } from '../../constants';
 import { useUserProfileContext } from '../../contexts';
-import { ProfileCardContent } from '../../elements';
 import { Route, Switch } from '../../router';
-import type { PropsOfComponent } from '../../styledSystem';
 import { ConnectedAccountsPage } from './ConnectedAccountsPage';
 import { DeletePage } from './DeletePage';
 import { EmailPage } from './EmailPage';
@@ -23,7 +21,7 @@ import { RootPage } from './RootPage';
 import { UsernamePage } from './UsernamePage';
 import { Web3Page } from './Web3Page';
 
-export const UserProfileRoutes = (props: PropsOfComponent<typeof ProfileCardContent>) => {
+export const UserProfileRoutes = () => {
   const { pages } = useUserProfileContext();
   const isAccountPageRoot =
     pages.routes[0].id === USER_PROFILE_NAVBAR_ROUTE_ID.ACCOUNT ||
@@ -46,114 +44,112 @@ export const UserProfileRoutes = (props: PropsOfComponent<typeof ProfileCardCont
   });
 
   return (
-    <ProfileCardContent contentRef={props.contentRef}>
-      <Switch>
-        {customPageRoutesWithContents}
-        <Route path={isAccountPageRoot ? undefined : 'account'}>
-          <Route
-            path='email-address'
-            flowStart
-          >
-            <Switch>
-              <Route path=':id/remove'>
-                <RemoveEmailPage />
-              </Route>
-              <Route path=':id'>
-                <EmailPage />
-              </Route>
-              <Route index>
-                <EmailPage />
-              </Route>
-            </Switch>
-          </Route>
-          <Route
-            path='phone-number'
-            flowStart
-          >
-            <Switch>
-              <Route path=':id/remove'>
-                <RemovePhonePage />
-              </Route>
-              <Route path=':id'>
-                <PhonePage />
-              </Route>
-              <Route index>
-                <PhonePage />
-              </Route>
-            </Switch>
-          </Route>
-          <Route
-            path='multi-factor'
-            flowStart
-          >
-            <Switch>
-              <Route path='totp/remove'>
-                <RemoveMfaTOTPPage />
-              </Route>
-              <Route path='backup_code/add'>
-                <MfaBackupCodeCreatePage />
-              </Route>
-              <Route path=':id/remove'>
-                <RemoveMfaPhoneCodePage />
-              </Route>
-              <Route path=':id'>
-                <MfaPage />
-              </Route>
-              <Route index>
-                <MfaPage />
-              </Route>
-            </Switch>
-          </Route>
-          <Route
-            path='connected-account'
-            flowStart
-          >
-            <Switch>
-              <Route path=':id/remove'>
-                <RemoveConnectedAccountPage />
-              </Route>
-              <Route index>
-                <ConnectedAccountsPage />
-              </Route>
-            </Switch>
-          </Route>
-          <Route
-            path='web3-wallet'
-            flowStart
-          >
-            <Switch>
-              <Route path=':id/remove'>
-                <RemoveWeb3WalletPage />
-              </Route>
-              <Route index>
-                <Web3Page />
-              </Route>
-            </Switch>
-          </Route>
-          <Route
-            path='username'
-            flowStart
-          >
-            <UsernamePage />
-          </Route>
-          {/*<Route path='security'>*/}
-          <Route
-            path='password'
-            flowStart
-          >
-            <PasswordPage />
-          </Route>
-          <Route
-            path='delete'
-            flowStart
-          >
-            <DeletePage />
-          </Route>
-          <Route index>
-            <RootPage />
-          </Route>
+    <Switch>
+      {customPageRoutesWithContents}
+      <Route path={isAccountPageRoot ? undefined : 'account'}>
+        <Route
+          path='email-address'
+          flowStart
+        >
+          <Switch>
+            <Route path=':id/remove'>
+              <RemoveEmailPage />
+            </Route>
+            <Route path=':id'>
+              <EmailPage />
+            </Route>
+            <Route index>
+              <EmailPage />
+            </Route>
+          </Switch>
         </Route>
-      </Switch>
-    </ProfileCardContent>
+        <Route
+          path='phone-number'
+          flowStart
+        >
+          <Switch>
+            <Route path=':id/remove'>
+              <RemovePhonePage />
+            </Route>
+            <Route path=':id'>
+              <PhonePage />
+            </Route>
+            <Route index>
+              <PhonePage />
+            </Route>
+          </Switch>
+        </Route>
+        <Route
+          path='multi-factor'
+          flowStart
+        >
+          <Switch>
+            <Route path='totp/remove'>
+              <RemoveMfaTOTPPage />
+            </Route>
+            <Route path='backup_code/add'>
+              <MfaBackupCodeCreatePage />
+            </Route>
+            <Route path=':id/remove'>
+              <RemoveMfaPhoneCodePage />
+            </Route>
+            <Route path=':id'>
+              <MfaPage />
+            </Route>
+            <Route index>
+              <MfaPage />
+            </Route>
+          </Switch>
+        </Route>
+        <Route
+          path='connected-account'
+          flowStart
+        >
+          <Switch>
+            <Route path=':id/remove'>
+              <RemoveConnectedAccountPage />
+            </Route>
+            <Route index>
+              <ConnectedAccountsPage />
+            </Route>
+          </Switch>
+        </Route>
+        <Route
+          path='web3-wallet'
+          flowStart
+        >
+          <Switch>
+            <Route path=':id/remove'>
+              <RemoveWeb3WalletPage />
+            </Route>
+            <Route index>
+              <Web3Page />
+            </Route>
+          </Switch>
+        </Route>
+        <Route
+          path='username'
+          flowStart
+        >
+          <UsernamePage />
+        </Route>
+        {/*<Route path='security'>*/}
+        <Route
+          path='password'
+          flowStart
+        >
+          <PasswordPage />
+        </Route>
+        <Route
+          path='delete'
+          flowStart
+        >
+          <DeletePage />
+        </Route>
+        <Route index>
+          <RootPage />
+        </Route>
+      </Route>
+    </Switch>
   );
 };
