+++ packages/clerk-js/src/ui/elements/Card/index.ts
@@ -1,0 +1,11 @@
+import { CardAlert } from './CardAlert';
+import { CardContent } from './CardContent';
+import { CardFooter } from './CardFooter';
+import { CardRoot } from './CardRoot';
+
+export const Card = {
+  Root: CardRoot,
+  Content: CardContent,
+  Footer: CardFooter,
+  Alert: CardAlert,
+};
