+++ packages/clerk-js/src/ui/elements/ProfileCard/index.ts
@@ -1,0 +1,7 @@
+import { ProfileCardContent } from './ProfileCardContent';
+import { ProfileCardRoot } from './ProfileCardRoot';
+
+export const ProfileCard = {
+  Root: ProfileCardRoot,
+  Content: ProfileCardContent,
+};
