+++ packages/clerk-js/src/ui/elements/index.ts
@@ -41,7 +41,6 @@ export * from './PersonalWorkspacePreview';
 export * from './Navbar';
 export * from './Breadcrumbs';
 export * from './FormContent';
-export * from './ProfileCardContent';
 export * from './IconButton';
 export * from './AvatarUploader';
 export * from './Actions';
@@ -57,3 +56,5 @@ export * from './Section';
 export * from './PreviewButton';
 export * from './InformationBox';
 export * from './withAvatarShimmer';
+export * from './Card';
+export * from './ProfileCard';
