+++ packages/clerk-js/src/ui/foundations/shadows.ts
@@ -4,7 +4,9 @@ export const shadows = Object.freeze({
   fabShadow: '0px 12px 24px rgba(0, 0, 0, 0.32)',
   buttonShadow:
     '0px 0px 0px 1px {{color}}, 0px 1px 1px 0px rgba(255, 255, 255, 0.07) inset, 0px 2px 3px 0px rgba(34, 42, 53, 0.20), 0px 1px 1px 0px rgba(0, 0, 0, 0.24)',
-  cardShadow:
+  cardRootShadow:
+    '0px 5px 15px 0px rgba(0, 0, 0, 0.08), 0px 15px 35px -5px rgba(25, 28, 33, 0.20), 0px 0px 0px 1px rgba(25, 28, 33, 0.06)',
+  cardContentShadow:
     '0px 0px 2px 0px rgba(0, 0, 0, 0.08), 0px 1px 2px 0px rgba(25, 28, 33, 0.06), 0px 0px 0px 1px rgba(25, 28, 33, 0.04)',
   actionCardShadow:
     '0px 1px 4px 0px rgba(0, 0, 0, 0.12), 0px 4px 8px 0px rgba(106, 115, 133, 0.12), 0px 0px 0px 1px rgba(106, 115, 133, 0.12)',
